// ClientApp/src/app/basic-setup/models/airCraftName.ts
export interface AirCraftName {
    airCraftNameId: number;
    departmentNameId:number;
    departmentName:string;
    name: string;
    image:string;
    overallLength: string;
    wingSpan: string;
    height: string;
    maxRange: string;
    endurance: string;
    maxTakeoffAndLandingWt: string;
    basicOperatingWt: string;
    cruisingSpeed: string;
    fuelCapacity: string;
    crew: string;
    madeBy: string;
    manufacturer: string;
    manufacturerMobile: string;
    aircraftStatus:number;
    email: string;
    remarks: string;
    status:boolean;
    isActive: boolean;
   
}